feat(auditoria): 20 issues prod - UX, copy, bugs criticos e seguranca
Auditoria completa easystok-web.fly.dev (batches 1-9):

Bugs criticos corrigidos:
- fix(routes): /pedidos/novo nao conflita mais com {id:guid}
- fix(historico): defaults 30d quando de/ate sao nulos (entradas + saidas)
- fix(lotes): LoteRepository inclui Itens no ListAsync (TotalUnidades=0 bug)
- fix(receita): DashboardAnalyticsQueries filtra apenas Natureza=Venda
- fix(velocity): PedidoEstoqueIntegrationService.DescontarAsync atualiza VelocidadeSaidaDiaria
- fix(concorrencia): DevolverAsync usa GetByIdComLockAsync + decimal qty + velocity update
- migration: backfill VelocidadeSaidaDiaria para dados historicos

Frontend / UX:
- refactor(sidebar): secoes Gestao/Vendas/Inteligencia/Sistema com nav-sep labels
- feat(pedidos): badges pagamento trinarios (Quitado/Parcial/Pendente)
  CTA banner amarelo em pedido entregue sem quitacao (!p.Quitado)
- feat(relatorios): datepickers substituem textarea JSON + submit.prevent fix
- feat(saidas): fallback precoReferencia||precoVendaSugerido||custoUnitario
- feat(analytics): alertas agrupados por ProdutoId, sort numerico por diasAteVencimento
- feat(estoque): sticky first column, aria-labels, copy (Giro/Parado/Critico/Todos/lotes)
- feat(dashboard): KPI cards clicareis, copy atualizado (Receita registrada, SKUs unicos)

Seguranca / qualidade:
- fix(xss): JavaScriptEncoder.Default.Encode em openModal e confirm() dialogs
- fix(analytics): OrderBy numerico em vez de lexicografico
- fix(copy): 50+ ajustes terminologia (PT-BR, empty states, toasts, capitalizacao)

+++ EasyStock.Api/Mobile/Services/MobileStockReconciler.cs
@@ -157,7 +157,7 @@ public class MobileStockReconciler(
             });
 
             // Espelha a quantidade no mobile_product (fonte da verdade local pro app)
-            mobileProduct.Stock = novo;
+            mobileProduct.Stock = (int)novo;
             mobileProduct.UpdatedAt = DateTime.UtcNow;
 
             // Marcador de divergência: registra MovimentacaoEstoque adicional
